Fix merge to main error
